rename: auth, customer domain 폴더 구조 변경
