+++ src/auth/application/jwt-refresh.strategy.ts
@@ -6,8 +6,8 @@ import {
 import { Injectable } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { Strategy, ExtractJwt } from 'passport-jwt';
-import { CustomerService } from 'src/customer/customer.service';
-import { Customer } from '../customer/entities/customer.entity';
+import { CustomerService } from 'src/customer/application/customer.service';
+import { Customer } from '../../customer/entities/customer.entity';
 
 @Injectable()
 export class JwtRefreshStrategy extends PassportStrategy(Strategy, 'refresh') {
