+++ src/auth/auth.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
-import { AuthService } from './auth.service';
-import { AuthController } from './auth.controller';
+import { AuthService } from './application/auth.service';
+import { AuthController } from './presentation/auth.controller';
 import { CustomerModule } from 'src/customer/customer.module';
 import { JwtModule } from '@nestjs/jwt';
 import { ConfigService } from '@nestjs/config';
-import { JwtAccessStrategy } from './jwt-access.strategy';
-import { JwtRefreshStrategy } from './jwt-refresh.strategy';
+import { JwtAccessStrategy } from './application/jwt-access.strategy';
+import { JwtRefreshStrategy } from './application/jwt-refresh.strategy';
 import { PassportModule } from '@nestjs/passport';
 import { CacheModule } from '../common/cache/cache.module';
