+++ src/auth/presentation/auth.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { AuthController } from './auth.controller';
-import { AuthService } from './auth.service';
+import { AuthService } from '../application/auth.service';
 
 describe('AuthController', () => {
   let controller: AuthController;
