+++ src/auth/presentation/auth.dto.ts
@@ -1,4 +1,4 @@
-import { CustomerDto } from 'src/customer/dto/customer.dto';
+import { CustomerDto } from 'src/customer/presentation/customer.dto';
 
 export class AuthDto extends CustomerDto {
   accessToken: string;
