+++ src/customer/application/customer.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { Customer } from './entities/customer.entity';
-import { CustomerDto } from './dto/customer.dto';
+import { Customer } from '../entities/customer.entity';
+import { CustomerDto } from '../presentation/customer.dto';
 
 @Injectable()
 export class CustomerService {
