+++ src/customer/customer.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { CustomerService } from './customer.service';
+import { CustomerService } from './application/customer.service';
 import { Customer } from './entities/customer.entity';
 
 @Module({
